# Conflicts:
#	app/Console/Kernel.php

+++ app/Helpers/Project/RevenuesKwhHelper.php
@@ -284,33 +284,63 @@ class RevenuesKwhHelper
             $distributionPartsKwh->parts_id = $revenuePartsKwh->id;
             $distributionPartsKwh->distribution_id = $distributionKwh->id;
             $distributionPartsKwh->status = 'concept';
+            $distributionPartsKwh->save();
         }
-        $distributionPartsKwh->save();
 
-        $dateBeginRevenues = Carbon::parse($revenuePartsKwh->revenuesKwh->date_begin)->format('Y-m-d');
-        list($quantityOfParticipationsAtStart, $quantityOfParticipations) = $this->determineParticipationsQuantityPart($dateBeginRevenues, $partDateBegin, $partDateEnd, $distributionPartsKwh);
-        $distributionPartsKwh->participations_quantity_at_start = $quantityOfParticipationsAtStart;
-        $distributionPartsKwh->participations_quantity = $quantityOfParticipations;
+        // Voor de zekerheid, we willen geen distributionPartsKwh records meer wijzigen qua aantallen die reeds status confirmed of processed hebben.
+        if(!in_array($distributionPartsKwh->status, ['confirmed', 'processed'])){
+            $dateBeginRevenues = Carbon::parse($revenuePartsKwh->revenuesKwh->date_begin)->format('Y-m-d');
+            list($quantityOfParticipationsAtStart, $quantityOfParticipations) = $this->determineParticipationsQuantityPart($dateBeginRevenues, $partDateBegin, $partDateEnd, $distributionPartsKwh);
+            $distributionPartsKwh->participations_quantity_at_start = $quantityOfParticipationsAtStart;
+            $distributionPartsKwh->participations_quantity = $quantityOfParticipations;
+            $distributionPartsKwh->delivered_kwh = 0;
 
-        $distributionPartsKwh->distributionKwh->newOrConceptDistributionValuesKwh()->where('parts_id', $revenuePartsKwh->id)->delete();
-        $this->saveDistributionValuesKwh($partDateBegin, $partDateEnd, $distributionPartsKwh);
+            $distributionPartsKwh->distributionKwh->newOrConceptDistributionValuesKwh()->where('parts_id', $revenuePartsKwh->id)->delete();
+            $this->saveDistributionValuesKwh($partDateBegin, $partDateEnd, $distributionPartsKwh);
+            $distributionPartsKwh->save();
+        }
 
-        $energySupplierUnknown = EnergySupplier::where('name', 'Onbekend')->first();
-        $distributionPartsKwh->delivered_kwh = 0;
-        if($addressEnergySupplier){
-            $distributionPartsKwh->es_id = $addressEnergySupplier ? $addressEnergySupplier->energy_supplier_id : null;
-            $distributionPartsKwh->energy_supplier_name = $addressEnergySupplier ? $addressEnergySupplier->energySupplier->name : null;
-            $distributionPartsKwh->energy_supplier_number = $addressEnergySupplier ? $addressEnergySupplier->es_number: null;
-            $distributionPartsKwh->is_visible = empty($distributionPartsKwh->remarks) ? false : true;
-        } else {
-            if($energySupplierUnknown){
-                $distributionPartsKwh->es_id = $energySupplierUnknown->id;
-                $distributionPartsKwh->energy_supplier_name = $energySupplierUnknown->name;
-                $distributionPartsKwh->energy_supplier_number = '';
-                $distributionPartsKwh->is_visible = 1;
+        // Address energy supplier niet meer wijzigen bij distributionPartsKwh als status processed is
+        if(!in_array($distributionPartsKwh->status, ['processed'])){
+            if(AddressEnergySupplier::where('address_id', $distributionPartsKwh->distributionKwh->participation->address_id)->where('energy_supplier_id', $distributionPartsKwh->es_id)->where('end_date', $distributionPartsKwh->partsKwh->date_end)->exists()){
+                $distributionPartsKwh->is_energy_supplier_switch = true;
+            } else {
+                $distributionPartsKwh->is_energy_supplier_switch = false;
+            }
+            if($distributionPartsKwh->distributionKwh->participation->date_terminated == $distributionPartsKwh->partsKwh->date_end){
+                $distributionPartsKwh->is_end_participation = true;
+            } else {
+                $distributionPartsKwh->is_end_participation = false;
+            }
+            if( $distributionPartsKwh->partsKwh->date_end && $distributionPartsKwh->partsKwh->date_end == $distributionPartsKwh->partsKwh->revenuesKwh->date_end ){
+                $distributionPartsKwh->is_end_total_period = true;
+            } else {
+                $distributionPartsKwh->is_end_total_period = false;
+            }
+            if( $distributionPartsKwh->partsKwh->date_end && Carbon::parse($distributionPartsKwh->partsKwh->date_end)->day == 31 && Carbon::parse($distributionPartsKwh->partsKwh->date_end)->month == 12 ){
+                $distributionPartsKwh->is_end_year_period = true;
+            } else {
+                $distributionPartsKwh->is_end_year_period = false;
+            }
+
+            if($addressEnergySupplier){
+                $distributionPartsKwh->es_id = $addressEnergySupplier ? $addressEnergySupplier->energy_supplier_id : null;
+                $distributionPartsKwh->energy_supplier_name = $addressEnergySupplier ? $addressEnergySupplier->energySupplier->name : null;
+                $distributionPartsKwh->energy_supplier_number = $addressEnergySupplier ? $addressEnergySupplier->es_number: null;
+//                $distributionPartsKwh->is_visible = empty($distributionPartsKwh->remarks) ? false : true;
+                $distributionPartsKwh->is_visible = $this->determineIsVisible($distributionPartsKwh);
+                $distributionPartsKwh->save();
+            } else {
+                $energySupplierUnknown = EnergySupplier::where('name', 'Onbekend')->first();
+                if($energySupplierUnknown){
+                    $distributionPartsKwh->es_id = $energySupplierUnknown->id;
+                    $distributionPartsKwh->energy_supplier_name = $energySupplierUnknown->name;
+                    $distributionPartsKwh->energy_supplier_number = '';
+                    $distributionPartsKwh->is_visible = true;
+                    $distributionPartsKwh->save();
+                }
             }
         }
-        $distributionPartsKwh->save();
     }
 
     /**
@@ -352,7 +382,7 @@ class RevenuesKwhHelper
             $dateEndForPeriod->endOfDay();
             $daysOfPeriod = $dateEndForPeriod->addDay()->diffInDays($dateBegin);
             if($dateBegin->format('Y-m-d') <= $dateEndMutation->format('Y-m-d')) {
-                RevenueDistributionValuesKwh::updateOrCreate(
+                RevenueDistributionValuesKwh::create(
                     [
                         'date_begin' => $dateBegin->format('Y-m-d'),
                         'date_end' => $dateEndMutation->format('Y-m-d'),
@@ -387,7 +417,7 @@ class RevenuesKwhHelper
                 'participations_quantity' => $participationsQuantity,
                 'quantity_multiply_by_days' => $participationsQuantity * $daysOfPeriod,
                 'delivered_kwh' => 0
-        ]);
+            ]);
     }
 
     public function checkRevenuePartsKwh(ParticipantProject $participant, $splitDate, AddressEnergySupplier $addressEnergySupplier = null)
@@ -725,12 +755,12 @@ class RevenuesKwhHelper
                 }
                 $totalDeliveredKwh = RevenueDistributionValuesKwh::where('revenue_id', $revenuePartsKwh->revenue_id)->where('distribution_id', $distributionPartsKwh->distribution_id)->where('parts_id', $revenuePartsKwh->id)->sum('delivered_kwh');
                 $distributionPartsKwh->delivered_kwh = $totalDeliveredKwh;
-                $distributionPartsKwh->is_visible = empty($distributionPartsKwh->remarks) ? false : true;
+                $distributionPartsKwh->is_visible = $this->determineIsVisible($distributionPartsKwh);
                 $distributionPartsKwh->save();
 
                 $newTotalDeliveredKwh = RevenueDistributionValuesKwh::where('revenue_id', $newRevenuePartsKwh->revenue_id)->where('distribution_id', $newDistributionPartsKwh->distribution_id)->where('parts_id', $newRevenuePartsKwh->id)->sum('delivered_kwh');
                 $newDistributionPartsKwh->delivered_kwh = $newTotalDeliveredKwh;
-                $newDistributionPartsKwh->is_visible = empty($newDistributionPartsKwh->remarks) ? false : true;
+                $newDistributionPartsKwh->is_visible = $this->determineIsVisible($newDistributionPartsKwh);
                 $newDistributionPartsKwh->save();
             }
             $revenuePartsKwhForRecalculate = RevenuePartsKwh::find($revenuePartsKwh->id);
@@ -833,6 +863,21 @@ class RevenuesKwhHelper
             ]);
         }
         return array($uniqueSplitDates, $periodParts);
+    }
+
+    /**
+     * @param RevenueDistributionPartsKwh $distributionPartsKwh
+     */
+    protected function determineIsVisible(RevenueDistributionPartsKwh $distributionPartsKwh): bool
+    {
+        if ($distributionPartsKwh->is_energy_supplier_switch
+            || $distributionPartsKwh->is_end_participation
+            || $distributionPartsKwh->is_end_total_period
+            || $distributionPartsKwh->is_end_year_period) {
+            return true;
+        } else {
+            return false;
+        }
     }
